requirements.txt
Django==4.2.7 
djangorestframework==3.14.0
pytest==7.4.3
djangorestframework-simplejwt==5.3.0
pytest-django==4.7.0 
dj-database-url==2.1.0 